arreglado proyecto base para modulo. solucion: añadido module al pom de gephi-plugins, añadido el servidor github con el token. +Info Taiga
