--- modules/columnCalculator/src/main/java/test.java
@@ -1,28 +1,0 @@
-
-
-public class Basic{
-
-	private int a;
-	private int b;
-
-	public Basic(int _a, int _b)
-	{
-		a = _a;
-		b = _b;
-	}
-
-	public int getA(){ return a;}
-	public int getB(){ return b;}
-	public void setA(int _a) { a = _a;}	
-	public void setA(int _b) { b = _b;}
-}
-
-public static void main {String args[]}
-{
-	Basic bas = new Basic(4,5);
-
-	int sum = getA() + getB();
-
-	System.out.println(sum);
-	
-}
